src/requirements.txt
playwright==1.42.0
langchain-community==0.0.25
langchain-core==0.1.28
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyyaml==6.0.1
fastapi==0.109.0
uvicorn==0.27.0
ollama==0.1.6